# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Football Training Board API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /teams:
    get:
      operationId: listTeams
      tags: [teams]
      summary: List all teams in the club
      responses:
        "200":
          description: List of teams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Team"
    post:
      operationId: createTeam
      tags: [teams]
      summary: Create a new team
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTeamRequest"
      responses:
        "201":
          description: Team created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /teams/{id}:
    get:
      operationId: getTeam
      tags: [teams]
      summary: Get a team by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Team details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
        "404":
          description: Team not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateTeam
      tags: [teams]
      summary: Update a team
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTeamRequest"
      responses:
        "200":
          description: Updated team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
    delete:
      operationId: deleteTeam
      tags: [teams]
      summary: Delete a team
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Team deleted
  /players:
    get:
      operationId: listPlayers
      tags: [players]
      summary: List all players in the club
      parameters:
        - name: teamId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of players
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Player"
    post:
      operationId: createPlayer
      tags: [players]
      summary: Create a new player
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlayerRequest"
      responses:
        "201":
          description: Player created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /players/{id}:
    get:
      operationId: getPlayer
      tags: [players]
      summary: Get a player by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Player details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
        "404":
          description: Player not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updatePlayer
      tags: [players]
      summary: Update a player
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePlayerRequest"
      responses:
        "200":
          description: Updated player
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
    delete:
      operationId: deletePlayer
      tags: [players]
      summary: Delete a player
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Player deleted
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Team:
      type: object
      properties:
        id:
          type: integer
        clubId:
          type: integer
        name:
          type: string
        ageGroup:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        coachId:
          type: ["integer", "null"]
        coachName:
          type: ["string", "null"]
        playerCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - clubId
        - name
        - playerCount
        - createdAt
    CreateTeamRequest:
      type: object
      properties:
        name:
          type: string
        ageGroup:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        coachId:
          type: ["integer", "null"]
      required:
        - name
    UpdateTeamRequest:
      type: object
      properties:
        name:
          type: string
        ageGroup:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        coachId:
          type: ["integer", "null"]
    Player:
      type: object
      properties:
        id:
          type: integer
        clubId:
          type: integer
        teamId:
          type: ["integer", "null"]
        teamName:
          type: ["string", "null"]
        firstName:
          type: string
        lastName:
          type: string
        dateOfBirth:
          type: ["string", "null"]
        nationality:
          type: ["string", "null"]
        position:
          type: ["string", "null"]
        jerseyNumber:
          type: ["integer", "null"]
        status:
          type: string
        height:
          type: ["integer", "null"]
        weight:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - clubId
        - firstName
        - lastName
        - status
        - createdAt
    CreatePlayerRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        teamId:
          type: ["integer", "null"]
        dateOfBirth:
          type: ["string", "null"]
        nationality:
          type: ["string", "null"]
        position:
          type: ["string", "null"]
        jerseyNumber:
          type: ["integer", "null"]
        status:
          type: string
        height:
          type: ["integer", "null"]
        weight:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
      required:
        - firstName
        - lastName
        - status
    UpdatePlayerRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        teamId:
          type: ["integer", "null"]
        dateOfBirth:
          type: ["string", "null"]
        nationality:
          type: ["string", "null"]
        position:
          type: ["string", "null"]
        jerseyNumber:
          type: ["integer", "null"]
        status:
          type: string
        height:
          type: ["integer", "null"]
        weight:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
...